crosspost_bot/requirements.txt
python-telegram-bot==20.7
vk-api==11.9.9
python-dotenv==1.0.0
requests==2.31.0
pillow==10.0.1